+++ tests/testthat/test-wk-crs.R
@@ -29,6 +29,11 @@ test_that("geodesic getting and setting works for wkb", {
   wk_is_geodesic(x) <- FALSE
   expect_false(wk_is_geodesic(x))
   expect_null(attr(x, "geodesic"))
+
+  wk_is_geodesic(x) <- wk_geodesic_inherit()
+  expect_identical(wk_is_geodesic(x), NA)
+
+  expect_error(wk_set_geodesic(x, "fish"), "must be TRUE, FALSE, or NA")
 })
 
 test_that("geodesic getting and setting works for wkt", {
@@ -39,10 +44,47 @@ test_that("geodesic getting and setting works for wkt", {
   wk_is_geodesic(x) <- FALSE
   expect_false(wk_is_geodesic(x))
   expect_null(attr(x, "geodesic"))
+
+  wk_is_geodesic(x) <- wk_geodesic_inherit()
+  expect_identical(wk_is_geodesic(x), NA)
+
+  expect_error(wk_set_geodesic(x, "fish"), "must be TRUE, FALSE, or NA")
 })
 
 test_that("geodesic setting gives a warning when this isn't supported", {
   expect_warning(wk_set_geodesic(xy(), TRUE), "for object of class 'wk_xy'")
+})
+
+test_that("wk_geodesic_output() works", {
+  expect_identical(
+    wk_is_geodesic_output(wkt(geodesic = FALSE), wkt(geodesic = FALSE)),
+    FALSE
+  )
+
+  expect_identical(
+    wk_is_geodesic_output(wkt(geodesic = TRUE), wkt(geodesic = TRUE)),
+    TRUE
+  )
+
+  expect_identical(
+    wk_is_geodesic_output(wkt(geodesic = wk_geodesic_inherit()), wkt(geodesic = FALSE)),
+    FALSE
+  )
+
+  expect_identical(
+    wk_is_geodesic_output(wkt(geodesic = FALSE), wkt(geodesic = wk_geodesic_inherit())),
+    FALSE
+  )
+
+  expect_error(
+    wk_is_geodesic_output(wkt(geodesic = TRUE), wkt(geodesic = FALSE)),
+    "differing values"
+  )
+
+  expect_error(
+    wk_is_geodesic_output(wkt(geodesic = FALSE), wkt(geodesic = TRUE)),
+    "differing values"
+  )
 })
 
 test_that("crs comparison works", {